%zd test

+++ tim.c
@@ -5,7 +5,7 @@ static int global_static;
 
 void f(int *p)
 {
-	printf("%d %d %d\n", global, global_static, *p);
+	printf("%zd %d %d\n", (size_t)global, global_static, *p);
 }
 
 int main()
